webapp - use arweave for storing and loading stories
IO interface is hardcoded to be arweave one and many of the functions for loading/saving content are implemented.
Geohashes are used (precisions from 1 to 10) when storing stories and when loading them.

For loading we use max precision that we can afford for visible region - I basically test increasing precision from 1 to 10 until geohashes covering the region become over 10 - then I use the set of geohashes from previous iteration.